requirements.txt
flask
twilio
pandas
ipykernel
matplotlib
kaggle
tensorflow
scikit-learn
ultralytics
opencv-python
pillow
flask-cors
